requirements.txt
anthropic
python-dotenv
requests
beautifulsoup4
